Add selected floor item information display
