+++ MicrowaveGame/Assets/Resources/Prefabs/Inventory/InventorySelectedFloorItemIndicator.prefab.meta
