+++ MicrowaveGame/Assets/Resources/Prefabs/Items/ItemSpeedIncrease.prefab
@@ -64,7 +64,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   Name: Speed Increase
-  Description: Increases movement speed by {0} for {1} seconds.
+  Description: Increases movement speed by {0}% for {1} second(s)
   IncreaseValue: 0.075
   DurationValue: 5
 --- !u!1001 &8036875327095706385
